Improve the regular expression checks for adding exercises
The clock and date validation expressions had lots of flaws because they were written when I was just starting with regex.
They should in theory be almost impossible to break now. The only thing to get wrong is February with 30 or 31 days...
The Uint rule was simplified and the Float rule was improved to not allow a leading dot with number behind.